Move stepper DAC to a separate file
